pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "chromadb>=1.3.5",
    "ddgs>=9.9.3",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.13.1",
    "google-genai>=1.54.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.2",
    "langchain-openai>=1.1.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "openai>=2.9.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tzdata>=2025.2",
]
